+++ src/core/pxtnOverDrive.h
@@ -3,6 +3,8 @@
 #ifndef pxtnOverDrive_H
 #define pxtnOverDrive_H
 
+#include "./pxTypedef.h"
+
 #include <pxwrDoc.h>
 
 #define TUNEOVERDRIVE_CUT_MAX       99.9f
@@ -17,13 +19,13 @@ class pxtnOverDrive
 {
 private:
 
-	bool  _b_played;
+	bool _b_played;
 
-	int   _group   ;
-	float _cut_f   ;
-	float _amp_f   ;
+	s32  _group   ;
+	f32  _cut_f   ;
+	f32  _amp_f   ;
 
-	int   _cut_16bit_top;
+	s32  _cut_16bit_top;
 public :
 
 
@@ -31,21 +33,21 @@ public :
 	~pxtnOverDrive();
 
 	void Tone_Ready();
-	void Tone_Supple( long *group_smps ) const;
+	void Tone_Supple( s32 *group_smps ) const;
 
 	bool Write( pxwrDoc *p_doc ) const;
 	bool Read ( pxwrDoc *p_doc, bool *pb_new_fmt );
 
 
-	float get_cut  ()const;
-	float get_amp  ()const;
-	int   get_group()const;
+	f32  get_cut  ()const;
+	f32  get_amp  ()const;
+	s32  get_group()const;
 
-	void  Set( float cut, float amp, int group );
+	void Set( f32 cut, f32 amp, s32 group );
 
-	bool  get_played()const;
-	void  set_played( bool b );
-	bool  switch_played();
+	bool get_played()const;
+	void set_played( bool b );
+	bool switch_played();
 
 };
