+++ src/core/pxtnService.h
@@ -2,6 +2,8 @@
 #define pxtnService_H
 
 
+#include "./pxTypedef.h"
+
 #include <pxwrDoc.h>
 #include "./pxPulse_NoiseBuilder.h"
 
@@ -84,12 +86,12 @@ enum pxtnERRORS
 
 typedef struct
 {
-	long  meas_start  ;
-	long  meas_end    ;
-	long  meas_repeat ;
-	long  start_sample; // for restart.
-	long  fadein_msec ;
-//	float volume      ; // 0.0 - 1.0
+	s32 meas_start  ;
+	s32 meas_end    ;
+	s32 meas_repeat ;
+	s32 start_sample; // for restart.
+	s32 fadein_msec ;
+//	f32 volume      ; // 0.0 - 1.0
 }
 pxtnVOMITPREPARATION;
 
@@ -110,33 +112,33 @@ private:
 
 	pxtnERRORS _err;
 
-	int _ch, _sps, _bps;
+	s32 _ch, _sps, _bps;
 
-	pxPulse_NoiseBuilder *_ptn_bldr;
+	pxPulse_NoiseBuilder* _ptn_bldr;
 
-	int _delay_max;	int _delay_num;	pxtnDelay     **_delays;
-	int _ovdrv_max;	int _ovdrv_num;	pxtnOverDrive **_ovdrvs;
-	int _woice_max;	int _woice_num;	pxtnWoice     **_woices;
-	int _unit_max ;	int _unit_num ;	pxtnUnit      **_units ;
+	s32 _delay_max;	s32 _delay_num;	pxtnDelay     **_delays;
+	s32 _ovdrv_max;	s32 _ovdrv_num;	pxtnOverDrive **_ovdrvs;
+	s32 _woice_max;	s32 _woice_num;	pxtnWoice     **_woices;
+	s32 _unit_max ;	s32 _unit_num ;	pxtnUnit      **_units ;
 
-	int        _group_num;
+	s32 _group_num;
 
-	bool _ReadVersion      ( pxwrDoc *p_doc, _enum_FMTVER *p_fmt_ver, unsigned short *p_exe_ver );
-	bool _ReadTuneItems    ( pxwrDoc *p_doc );
-	bool _x1x_Project_Read ( pxwrDoc *p_doc );
+	bool _ReadVersion      ( pxwrDoc* p_doc, _enum_FMTVER* p_fmt_ver, u16 *p_exe_ver );
+	bool _ReadTuneItems    ( pxwrDoc* p_doc );
+	bool _x1x_Project_Read ( pxwrDoc* p_doc );
 
-	bool _io_Read_Delay    ( pxwrDoc *p_doc, bool *pb_new_fmt );
-	bool _io_Read_OverDrive( pxwrDoc *p_doc, bool *pb_new_fmt );
-	bool _io_Read_Woice    ( pxwrDoc *p_doc, bool *pb_new_fmt, pxtnWOICETYPE type );
-	bool _io_Read_OldUnit  ( pxwrDoc *p_doc, bool *pb_new_fmt, int ver        );
+	bool _io_Read_Delay    ( pxwrDoc* p_doc, bool* pb_new_fmt );
+	bool _io_Read_OverDrive( pxwrDoc* p_doc, bool* pb_new_fmt );
+	bool _io_Read_Woice    ( pxwrDoc* p_doc, bool* pb_new_fmt, pxtnWOICETYPE type );
+	bool _io_Read_OldUnit  ( pxwrDoc* p_doc, bool* pb_new_fmt, s32 ver            );
 
-	bool _io_assiWOIC_w    ( pxwrDoc *p_doc, int idx          ) const;
-	bool _io_assiWOIC_r    ( pxwrDoc *p_doc, bool *pb_new_fmt );
-	bool _io_assiUNIT_w    ( pxwrDoc *p_doc, int idx          ) const;
-	bool _io_assiUNIT_r    ( pxwrDoc *p_doc, bool *pb_new_fmt );
+	bool _io_assiWOIC_w    ( pxwrDoc* p_doc, s32 idx          ) const;
+	bool _io_assiWOIC_r    ( pxwrDoc* p_doc, bool* pb_new_fmt );
+	bool _io_assiUNIT_w    ( pxwrDoc* p_doc, s32 idx          ) const;
+	bool _io_assiUNIT_r    ( pxwrDoc* p_doc, bool* pb_new_fmt );
 
-	bool _io_UNIT_num_w    ( pxwrDoc *p_doc ) const;
-	int  _io_UNIT_num_r    ( pxwrDoc *p_doc, bool *pb_new_fmt );
+	bool _io_UNIT_num_w    ( pxwrDoc* p_doc ) const;
+	s32  _io_UNIT_num_r    ( pxwrDoc* p_doc, bool *pb_new_fmt );
 
 	bool _x3x_CorrectKeyEvent();
 	bool _x3x_AddCorrectEvent();
@@ -145,121 +147,121 @@ private:
 	//////////////
 	// Vomit..
 	//////////////
-	bool  _vmt_b_ready   ;
-	bool  _vmt_b_init    ;
+	bool _vmt_b_ready   ;
+	bool _vmt_b_init    ;
 
-	bool  _vmt_b_mute    ;
-	bool  _vmt_b_loop    ;
+	bool _vmt_b_mute    ;
+	bool _vmt_b_loop    ;
 
-	long  _vmt_smp_smooth;
-	float _vmt_clock_rate; // as the sample
-	long  _vmt_smp_count ;
-	long  _vmt_smp_start ;
-	long  _vmt_smp_end   ;
-	long  _vmt_smp_repeat;
+	s32  _vmt_smp_smooth;
+	f32  _vmt_clock_rate; // as the sample
+	s32  _vmt_smp_count ;
+	s32  _vmt_smp_start ;
+	s32  _vmt_smp_end   ;
+	s32  _vmt_smp_repeat;
 		
-	long  _vmt_fade_count;
-	long  _vmt_fade_max  ;
-	long  _vmt_fade_fade ;
-	float _vmt_master_vol;
+	s32  _vmt_fade_count;
+	s32  _vmt_fade_max  ;
+	s32  _vmt_fade_fade ;
+	f32  _vmt_master_vol;
 		
-	int   _vmt_top;
-	long  _vmt_smp_skip   ;
-	long  _vmt_time_pan_index;
+	s32  _vmt_top       ;
+	s32  _vmt_smp_skip  ;
+	s32  _vmt_time_pan_index;
 
-	float _vmt_bt_tempo;
+	f32  _vmt_bt_tempo  ;
 
 	// for make now-meas
-	long  _vmt_bt_clock;
-	long  _vmt_bt_num  ;
+	s32  _vmt_bt_clock  ;
+	s32  _vmt_bt_num    ;
 
-	long  *_vmt_group_smps;
+	s32* _vmt_group_smps;
 
 
-	const EVERECORD   *_vmt_p_eve;
+	const EVERECORD*   _vmt_p_eve;
 
-	pxPulse_Frequency *_vmt_freq ;
+	pxPulse_Frequency* _vmt_freq ;
 
 	void _vmt_constructor();
 	void _vmt_destructer ();
 	bool _vmt_init       ();
 
 
-	void _vmt_ResetVoiceOn( pxtnUnit *p_u, long w ) const;
+	void _vmt_ResetVoiceOn( pxtnUnit* p_u, s32 w ) const;
 	void _vmt_InitUnitTone();
-	bool _vmt_PXTONE_SAMPLE( void *p_data );
+	bool _vmt_PXTONE_SAMPLE( void* p_data );
 
 public :
 
 	 pxtnService();
 	~pxtnService();
 
-	pxtnText    *text  ;
-	pxtnMaster  *master;
-	pxtnEvelist *evels ;
+	pxtnText*    text  ;
+	pxtnMaster*  master;
+	pxtnEvelist* evels ;
 
 	bool Init         ( bool b_reserve_evels ); // ptc reserves evels.
 	void Clear        ( bool b_release_evels ); // ptc doesn't release evels.
 
-	bool Save         ( pxwrDoc *p_doc, bool bTune, unsigned short exe_ver );
-	long PreCountEvent( pxwrDoc *p_doc, long max_unit      );
-	bool Read         ( pxwrDoc *p_doc, bool bPermitPTTUNE );
+	bool Save         ( pxwrDoc* p_doc, bool bTune, u16 exe_ver );
+	s32  PreCountEvent( pxwrDoc* p_doc, s32 max_unit            );
+	bool Read         ( pxwrDoc* p_doc, bool bPermitPTTUNE      );
 	void AdjustMeasNum();
 
-//	const char *GetIOError();
+//	const char* GetIOError();
 
-	const char *get_last_error_text() const;
+	const char* get_last_error_text() const;
 
 	bool Tone_Ready();
 	void Tone_Clear();
 
-	int  Group_Num() const;
+	s32  Group_Num() const;
 
 	// delay.
-	int  Delay_Num() const;
-	int  Delay_Max() const;
-	bool Delay_Set         ( int idx, DELAYUNIT unit, float freq, float rate, int group );
-	bool Delay_Add         (          DELAYUNIT unit, float freq, float rate, int group );
-	bool Delay_Remove      ( int idx );
-	bool Delay_ReadyTone   ( int idx );
-	pxtnDelay *Delay_Get   ( int idx );
+	s32  Delay_Num() const;
+	s32  Delay_Max() const;
+	bool Delay_Set         ( s32 idx, DELAYUNIT unit, f32 freq, f32 rate, s32 group );
+	bool Delay_Add         (          DELAYUNIT unit, f32 freq, f32 rate, s32 group );
+	bool Delay_Remove      ( s32 idx );
+	bool Delay_ReadyTone   ( s32 idx );
+	pxtnDelay* Delay_Get   ( s32 idx );
 
 
 	// over drive.
-	int  OverDrive_Num() const;
-	int  OverDrive_Max() const;
-	bool OverDrive_Set          ( int idx, float cut, float amp, int group );
-	bool OverDrive_Add          (          float cut, float amp, int group );
-	bool OverDrive_Remove       ( int idx );
-	void OverDrive_ReadyTone    ( int idx );
-	pxtnOverDrive *OverDrive_Get( int idx );
+	s32  OverDrive_Num() const;
+	s32  OverDrive_Max() const;
+	bool OverDrive_Set          ( s32 idx, f32 cut, f32 amp, s32 group );
+	bool OverDrive_Add          (          f32 cut, f32 amp, s32 group );
+	bool OverDrive_Remove       ( s32 idx );
+	void OverDrive_ReadyTone    ( s32 idx );
+	pxtnOverDrive* OverDrive_Get( s32 idx );
 
 	// woice.
-	int  Woice_Num() const;
-	int  Woice_Max() const;
-	const pxtnWoice *Woice_Get( int idx ) const;
-	pxtnWoice       *Woice_Get_variable( int idx );
+	s32  Woice_Num() const;
+	s32  Woice_Max() const;
+	const pxtnWoice* Woice_Get( s32 idx ) const;
+	pxtnWoice*       Woice_Get_variable( s32 idx );
 
-	bool Woice_Load     ( int idx, const char *path, pxtnWOICETYPE type, bool *pb_new_fmt );
-	bool Woice_ReadyTone( int idx );
-	void Woice_Remove   ( int idx );
-	void Woice_Replace  ( int old_place, int new_place );
+	bool Woice_Load     ( s32 idx, const char *path, pxtnWOICETYPE type, bool *pb_new_fmt );
+	bool Woice_ReadyTone( s32 idx );
+	void Woice_Remove   ( s32 idx );
+	void Woice_Replace  ( s32 old_place, s32 new_place );
 
 	// unit.
-	int   Unit_Num() const;
-	int   Unit_Max() const;
-	const pxtnUnit *Unit_Get         ( int idx ) const;
-	pxtnUnit       *Unit_Get_variable( int idx );
+	s32   Unit_Num() const;
+	s32   Unit_Max() const;
+	const pxtnUnit* Unit_Get         ( s32 idx ) const;
+	pxtnUnit*       Unit_Get_variable( s32 idx );
 
-	void Unit_Remove   ( int idx );
-	void Unit_Replace  ( int old_place, int new_place );
+	void Unit_Remove   ( s32 idx );
+	void Unit_Replace  ( s32 old_place, s32 new_place );
 	bool Unit_AddNew   ();
 	void Unit_SetOpratedAll( bool b );
-	void Unit_Solo( int idx );
+	void Unit_Solo( s32 idx );
 
 	// q
-	void Quality_Set( int    ch, int    sps, int    bps );
-	void Quality_Get( int *p_ch, int *p_sps, int *p_bps ) const;
+	void Quality_Set( s32    ch, s32    sps, s32    bps );
+	void Quality_Get( s32 *p_ch, s32 *p_sps, s32 *p_bps ) const;
 
 
 	//////////////
@@ -270,18 +272,18 @@ public :
 
 	void vmt_set_mute( bool b );
 	void vmt_set_loop( bool b );
-	void vmt_set_fade( long fade, long msec );
-	void vmt_set_master_volume( float v );
+	void vmt_set_fade( s32 fade, s32 msec );
+	void vmt_set_master_volume( f32 v );
 
-	int  vmt_get_now_clock      () const;
-	int  vmt_get_sampling_offset() const;
+	s32  vmt_get_now_clock      () const;
+	s32  vmt_get_sampling_offset() const;
 
 	bool vmt_preparation( const pxtnVOMITPREPARATION *p_build );
 
-	bool Vomit( void* p_buf, long size );
+	bool Vomit( void* p_buf, s32 size );
 };
 
-int pxtnService_vmt_CalcSampleNum( int meas_num, int beat_num, int sps, float beat_tempo );
+s32 pxtnService_vmt_CalcSampleNum( s32 meas_num, s32 beat_num, s32 sps, f32 beat_tempo );
 
 
 #endif
